ARM: dts: qcom: msm8226-motorola-falcon: specify vddio_disp output voltage
After some digging in downstream sources, it was found that the vddio_disp
regulator's output voltage is 1.8V. This is further confirmed by the
troubleshooting guide. Specify its output voltage as such.
While at it, add a comment specifying the IC, which according to the
schematic is the TI TPS22902.

Signed-off-by: Stanislav Jakubek <stano.jakubek@gmail.com>
Reviewed-by: Konrad Dybcio <konrad.dybcio@oss.qualcomm.com>
Link: https://lore.kernel.org/r/6143603464a65aebbed281fe6c6164316dd07269.1746711762.git.stano.jakubek@gmail.com
Signed-off-by: Bjorn Andersson <andersson@kernel.org>

[ upstream commit: e41ece7cd173486f6767de050deb2d8f0882a82a ]