dt-bindings: drm/bridge: ti-sn65dsi83: drop $ref to fix lvds-vod* warnings
The kernel test robot reported a warning related to the use of "$ref"
type definitions for custom endpoint properties
- "ti,lvds-vod-swing-clock-microvolt" and
- "ti,lvds-vod-swing-data-microvolt".

Using "$ref" with "uint32-array" is not correctly handled in this
context. Removing "$ref" and relying solely on "maxItems: 2" enforces
the intended requirement of specifying exactly two values, without
triggering a schema validation warning.

Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202505021937.efnQPPqx-lkp@intel.com/
Signed-off-by: Andrej Picej <andrej.picej@norik.com>
Link: https://lore.kernel.org/r/20250529053654.1754926-1-andrej.picej@norik.com
Signed-off-by: Rob Herring (Arm) <robh@kernel.org>

[ upstream commit: 400123bd0107175e92a9780b97f7a5934eb0a991 ]