ARM: dts: imx6ulz-bsh-smm-m2: fix resume via console
Despite the current configuration being compliant with the technical
reference manual (TRM), testing on the system showed that resuming from
suspend via UART4 (used as the console) fails unless any other UART
is also enabled.

In our use case, UART2 is enabled to ensure reliable resume when UART4
is used as the console.

Signed-off-by: Wolfgang Birkner <wolfgang.birkner@bshg.com>
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Acked-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 9db115e231a665e82429ea63e2efda4abd4390fe ]