arm64: zynqmp: Fix pwm-fan polarity
The correct operating mode for the fan is inversed (1). The previous pwm
driver implementation had a bug and the polarity information was propagated
incorrectly to the kernel. The normal (0) polarity specified in the device
tree was incorrectly clearing the polarity bit in the counter control
register. After the bug fix, setting the polarity to inversed (1) in the
device tree will clear the polarity bit.
Also provide label for custom description based on this SOM.

Signed-off-by: Vishal Patel <vishal.patel@amd.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>
Link: https://lore.kernel.org/r/e6dc9ce1e0b97c9d9b073c323cc94a5d7a88b94a.1752835502.git.michal.simek@amd.com

[ upstream commit: 7cf18a364ed2d56cf44769d1a3690aa0bf4d5e9e ]