arm64: dts: mediatek: mt8195-cherry: Add missing regulators to rt5682
Add the missing DBVDD and LDO1-IN power supplies to the codec
node as both RT5682i and RT5682s require those.

This commit only fixes a dtbs_check warning but doesn't produce
any functional changes because the VIO18 LDO is already powered
on because it's assigned as AVDD supply anyway.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Link: https://lore.kernel.org/r/20250724083914.61351-37-angelogioacchino.delregno@collabora.com
Signed-off-by: Matthias Brugger <matthias.bgg@gmail.com>

[ upstream commit: 45049abe5bcda5b049d22423f10f74e38041a435 ]