dt-bindings: i2c: apple,i2c: Add apple,t6020-i2c compatible
After discussion with the devicetree maintainers we agreed to not extend
lists with the generic compatible "apple,i2c" anymore [1]. Use
"apple,t8103-i2c" as fallback compatible as it is the SoC the driver
and bindings were written for.

This block is compatible with t8103, so just add the new per-SoC
compatible using apple,t8103-i2c as base.

[1]: https://lore.kernel.org/asahi/12ab93b7-1fc2-4ce0-926e-c8141cfe81bf@kernel.org/

Signed-off-by: Janne Grunau <j@jannau.net>
Acked-by: Andi Shyti <andi.shyti@kernel.org>
Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>

[ upstream commit: f77a13df57e803fe574dd29c5454088fcc6f3285 ]